promote beta endpoints to v4
